requirements.txt
matrix-nio[e2e]==0.24.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
aiofiles==23.2.1
aiohttp==3.9.1
